pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beam_sup"
version = "0.1.0"
description = "A Python package for analyzing BEAM output"
authors = [
    {name = "Stephen Staklinski", email = "staklins@cshl.edu"},
]
dependencies = [
    "black",
    "ete3==3.1.3",
    "pandas>=2.0.0",
    "dendropy>=4.5.2",
    "numpy>=1.19.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.5.0",
    "networkx>=3.0",
    "pydot>=1.4.2",
    "PyQt5==5.15.9",
    "cassiopeia-lineage @ git+https://github.com/YosefLab/Cassiopeia@master",
]
requires-python = ">=3.8,<=3.12"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
run_met_cancer_barcode_simulation = "beam_sup.beam_sup:run_simulation"

[project.urls]
Homepage = "https://github.com/StephenStaklinski/beam_sup"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=beam_sup --cov-report=term-missing" 